# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boleteria Online",
    "description": "Colleción de API's correspondientes al sistema de Pasajes Online.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@boleteriasystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/destinos": {
      "get": {
        "tags": [
          "Destinos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseICollectionWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseICollectionWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseICollectionWebResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Destinos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Destinos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/destinos/{id}": {
      "get": {
        "tags": [
          "Destinos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Destinos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DestinoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinoResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DestinoRequest": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DestinoResponse": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DestinoResponseICollectionWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DestinoResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DestinoResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/DestinoResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}